Split database code a bit more
